pyproject.toml
[project]
name = "threads-diary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "marimo>=0.11.13",
    "polars>=1.23.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.9",
]
